Changes 2025 part 6- Changes from Gerald Following changes has been done ,
Changes
1.While we’re making updates to the Estimator, the Recruiters have asked that we reduce the number of required fields on the
 Personal tab. So, please remove the “required” status for Telephone, Email, State, and Zip Code (keep it in First Name, Last Name, and Date of Birth).-  Done - removed the mandatory flag
2.One of the recruiters has informed me that when entering the $ 1,000 Freshmen Registration Bonus on the Institutional Tab (from the drop-down in Scholarship #7) 
an error message pops up “ Scholarship #7 Amount is over the limit.”  (see screenshot) 
It should not do this!
It appears that this error occurs regardless of the drop-down value selected.  I did not experience this error during testing, but it may be something I missed. I’m not sure.- This similar to point 4  amount due calculation is fixed now.
3.SAI changes - fixed the bug and altered the sai limits 
4.Due amount reconcilation -done if financial aid is more than maxaid we always consider financial aid.

NOTE:Please Change the private.properties,build.properties under nbproject